Refactored provider.
